+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `.expression.Operator#IN`. Example use: `web.request.remote.address in io.github.ratelimiter.Bots.getBotIps()`.
+- `Bandwidths#getPermitsPer(TimeUnit)`.
 
 ### Removed